requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI & LangChain
langchain==0.0.335
langchain-google-genai==0.0.8
google-generativeai==0.3.2

# File Processing
PyMuPDF==1.23.8
python-docx==1.1.0
Pillow==10.1.0

# Additional utilities
aiofiles==23.2.1